requirements.txt
torch
numpy
pandas
transformers
tqdm
scikit-learn
gensim
matplotlib
seaborn
nltk 
jieba
python-docx